SGA-Backend/requirements.txt
fastapi
sqlalchemy
passlib
uvicorn
python-dotenv
pyodbc
bcrypt
python-multipart